[UPDATE] 이미지 삭제시 null 체크

+++ src/main/java/com/huitdduru/madduru/mypage/service/MyPageServiceImpl.java
@@ -45,7 +45,8 @@ public class MyPageServiceImpl implements MyPageService {
 
         String imagePath = image != null ? fileUploader.uploadFile(image, uuid) : null;
 
-        fileUploader.removeFile(user.getImagePath());
+        if(user.getImagePath() != null)
+            fileUploader.removeFile(user.getImagePath());
 
         userRepository.save(user.setImagePath(imagePath));
     }
